pip_requirements.txt
flask-login==0.6.*
bcrypt==4.0.*